Move tests to a Tests folder and add check prefix
- Rename `UnparserTests` directory to `Tests` since we can use its tests
  as input to both the unparser and taint analysis tests.
- Add check-prefx optino to all tests to specify which inputs/outputs to
  use for testing specific tools.
- Add taint analysis tests to assgn_p_san.c
