+++ test/Tests/if.c
@@ -1,0 +1,31 @@
+// RUN: unparse %s | clang-format | FileCheck %s --color --check-prefix=UNPARSE
+
+#include <stdio.h>
+
+void sanitize(int *x) {
+        if (*x != 0) {
+                *x = 0;
+        } else {
+        }
+}
+
+int main(void) {
+        int x;
+        scanf("%d", &x);
+        sanitize(&x);
+
+        return 0;
+}
+
+// UNPARSE: void sanitize(int *x) {
+// UNPARSE:         if (*x != 0) {
+// UNPARSE:                 *x = 0;
+// UNPARSE:         } else {
+// UNPARSE:         }
+// UNPARSE: }
+// UNPARSE: int main(void) {
+// UNPARSE:         int x;
+// UNPARSE:         scanf("%d", &x);
+// UNPARSE:         sanitize(&x);
+// UNPARSE:         return 0;
+// UNPARSE: }
